requirements.txt
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
PyGObject>=3.44.1
python-dateutil>=2.8.2
requests>=2.31.0
oauth2client>=4.1.3
cachetools>=5.3.0 